backend/zenexotics_backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
awsebcli==3.24
bcrypt==4.3.0
billiard==4.2.1
blessed==1.21.0
boto3==1.37.38
botocore==1.37.38
celery==5.5.3
cement==2.10.14
certifi==2025.4.26
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.3
daphne==4.0.0
decorator==5.2.1
Deprecated==1.2.18
dj-database-url==2.3.0
Django==4.2.7
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dotenv==0.9.9
fabric==3.2.2
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
invoke==2.2.0
jmespath==1.0.1
kombu==5.5.4
msgpack==1.1.0
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pillow==10.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
s3transfer==0.11.5
semantic-version==2.10.0
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.4.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
watchdog==4.0.2
wcwidth==0.2.13
whitenoise==6.6.0
wrapt==1.17.2
zope.interface==7.2
